The last carton could be sold to two people at once
Phase 8 — Sales. Four documents, because a sale is four separate moments:
the customer asks, the goods leave, the money becomes owed, the money
arrives. Tying any two together forces the wrong thing to wait.

Confirming an order finally gives Inventory's Reserved state a writer.
Until now four states were declared and only three were ever written.
Held stock stays on the shelf — taking it off would mean someone counting
the shelf finds more than the books say, with nothing wrong. Not holding
it at all means the same last carton gets sold twice, and that is
discovered in front of the customer.

The challan takes goods off the floor and releases the order hold in a
single movement row. Two rows would eventually mean one written and one
not, and then the goods read as both gone and still reserved — Available
dropping twice for one event. Nothing reaches the ledger: goods leaving
is not a sale, they can come back, and the price is not settled.

The invoice posts income, the receivable and the cost of goods together.
Without the cost, the profit and loss shows revenue with nothing behind
it and inventory that was sold stays on the balance sheet as an asset.
The cost is frozen at invoice time, deliberately breaking this codebase's
own rule that figures are counted from the ledger — purchase prices move,
and last month's profit must not move with them.

A counter sale has no challan, so there the invoice is what moves stock.

Collections allocate against invoices rather than sitting as a lump, so
"which invoice is still open" has an answer and chasing names the right
document.

Found by driving the forms, not by tests:
- A legitimate challan was rejected as belonging to another customer.
  The form posts an id as a string, the database returns an int, and the
  guard used strict !==. Tests passed because they pass ints. The same
  latent comparison existed in four other places, including Purchase.
- The demo seeder had no customers, so every sales screen opened onto an
  empty dropdown — the same gap suppliers had.

And one that had been wrong from the start: the salesman role was granted
`sales.% OR customer.%` with an exclusion appended, but AND binds tighter
than OR, so the exclusion only ever applied to the second prefix.
Salesmen have always had customer.credit_limit.override despite the
comment beside it saying they must not, and the new
sales.discount.override went the same way. Both prefixes are now grouped.

+++ app/Modules/Purchase/Services/PurchaseBillService.php
@@ -380,11 +380,11 @@ final class PurchaseBillService
 
         // অন্য কোম্পানির চালানের id পাঠিয়ে দেওয়া আটকায় — গ্লোবাল স্কোপ
         // সন্তান-টেবিলে নেই, বাবার উপর আছে
-        if ($receiptLine->receipt->company_id !== CompanyContext::id()) {
+        if ((int) $receiptLine->receipt->company_id !== (int) CompanyContext::id()) {
             throw ValidationException::withMessages(['lines' => __('purchase::validation.unknown_receipt_line')]);
         }
 
-        if ($receiptLine->receipt->supplier_id !== $bill->supplier_id) {
+        if ((int) $receiptLine->receipt->supplier_id !== (int) $bill->supplier_id) {
             throw ValidationException::withMessages(['lines' => __('purchase::validation.receipt_other_supplier')]);
         }
 
@@ -396,7 +396,7 @@ final class PurchaseBillService
             ]);
         }
 
-        if ($receiptLine->product_id !== $productId) {
+        if ((int) $receiptLine->product_id !== $productId) {
             throw ValidationException::withMessages(['lines' => __('purchase::validation.line_product_mismatch')]);
         }
